+++ src/main/java/net/landofrails/landofsignals/contentpacks/ContentPackHandlerV2.java
@@ -63,7 +63,7 @@ public class ContentPackHandlerV2 {
             contentPack.getContentSets().forEach(path -> loadSet(zip, path, contentPack, isUTF8));
         }
         if (!hasContent && !hasContentSets) {
-            ModCore.warn("ContentPack %s does not contain any blocks");
+            ModCore.warn("ContentPack %s does not contain any blocks", contentPack.getId());
         }
     }
 
@@ -207,7 +207,7 @@ public class ContentPackHandlerV2 {
                             loadSignal(zip, entryPath, contentPack, isUTF8);
                             break;
                         case BLOCKCOMPLEXSIGNAL:
-                            loadComplexSignal(zip, path, contentPack, isUTF8);
+                            loadComplexSignal(zip, entryPath, contentPack, isUTF8);
                             break;
                         case BLOCKSIGN:
                             loadSign(zip, entryPath, contentPack, isUTF8);