+++ src/main/java/net/landofrails/landofsignals/render/block/TileComplexSignalRender.java
@@ -14,6 +14,7 @@ import net.landofrails.api.contentpacks.v2.parent.ContentPackModel;
 import net.landofrails.landofsignals.LOSBlocks;
 import net.landofrails.landofsignals.LandOfSignals;
 import net.landofrails.landofsignals.tile.TileComplexSignal;
+import net.landofrails.landofsignals.utils.HighlightingUtil;
 import net.landofrails.landofsignals.utils.Static;
 import org.lwjgl.opengl.GL11;
 
@@ -30,7 +31,7 @@ public class TileComplexSignalRender {
     private static final Map<String, OBJRender> cache = new HashMap<>();
     private static final Map<String, List<String>> groupCache = new HashMap<>();
 
-    private static void checkCache(String blockId, Collection<ContentPackSignalGroup> groups, String identifier) {
+    public static void checkCache(String blockId, Collection<ContentPackSignalGroup> groups, String identifier) {
 
         // Get first group, get first state, get first model
         Optional<String> firstPath = groups.iterator().next().getStates().values().iterator().next().getModels().keySet().stream().findFirst();
@@ -50,7 +51,7 @@ public class TileComplexSignalRender {
 
     }
 
-    private static void checkCache(String blockId, Map<String, ContentPackModel[]> models, String identifier, boolean checkIfAlreadyExisting) {
+    public static void checkCache(String blockId, Map<String, ContentPackModel[]> models, String identifier, boolean checkIfAlreadyExisting) {
         if (checkIfAlreadyExisting) {
             Optional<String> firstPath = models.keySet().stream().findFirst();
             if (!firstPath.isPresent())
@@ -74,7 +75,7 @@ public class TileComplexSignalRender {
                 for (ContentPackModel signalModel : modelEntry.getValue()) {
                     String[] groups = signalModel.getObj_groups();
                     if (groups.length > 0) {
-                        Predicate<String> targetGroup = renderOBJGroup -> Arrays.stream(groups).anyMatch(renderOBJGroup::startsWith);
+                        Predicate<String> targetGroup = renderOBJGroup -> Arrays.stream(groups).filter(Objects::nonNull).anyMatch(renderOBJGroup::startsWith);
                         List<String> modes = renderer.model.groups().stream().filter(targetGroup)
                                 .collect(Collectors.toCollection(ArrayList::new));
                         String groupCacheId = objId + "@" + String.join("+", groups);
@@ -104,12 +105,17 @@ public class TileComplexSignalRender {
         renderBase(id, tsp);
         renderSignals(id, tsp);
 
+        if(tsp.isHighlighting()){
+            HighlightingUtil.renderHighlighting();
+        }
+
     }
 
     @SuppressWarnings("java:S1134")
     private static void renderBase(String blockId, TileComplexSignal tile) {
 
         final Vec3d offset = tile.getOffset();
+        final Vec3d customScaling = tile.getScaling();
 
         checkCache(blockId, LOSBlocks.BLOCK_COMPLEX_SIGNAL.getContentpackComplexSignals().get(blockId).getBase(), "/base/", true);
 
@@ -124,7 +130,7 @@ public class TileComplexSignalRender {
             for (ContentPackModel baseModel : baseModels.getValue()) {
                 final ContentPackBlock block = baseModel.getBlock();
                 final Vec3d translate = block.getAsVec3d(block::getTranslation).add(offset);
-                final Vec3d scale = block.getAsVec3d(block::getScaling);
+                final Vec3d scale = Static.multiply(block.getAsVec3d(block::getScaling), customScaling);
                 final Vec3d rotation = block.getAsVec3d(block::getRotation);
                 try (OpenGL.With ignored1 = OpenGL.matrix(); OpenGL.With ignored2 = renderer.bindTexture(baseModel.getTextures())) {
 
@@ -157,6 +163,7 @@ public class TileComplexSignalRender {
     private static void renderSignals(final String blockId, final TileComplexSignal tile) {
 
         final Vec3d offset = tile.getOffset();
+        final Vec3d customScaling = tile.getScaling();
 
         final Map<String, ContentPackSignalGroup> signalGroups = LOSBlocks.BLOCK_COMPLEX_SIGNAL.getContentpackComplexSignals().get(blockId).getSignals();
 
@@ -183,7 +190,7 @@ public class TileComplexSignalRender {
                 for (ContentPackModel signalModel : signalModels.getValue()) {
                     ContentPackBlock block = signalModel.getBlock();
                     final Vec3d translate = block.getAsVec3d(block::getTranslation).add(offset);
-                    final Vec3d scale = block.getAsVec3d(block::getScaling);
+                    final Vec3d scale = Static.multiply(block.getAsVec3d(block::getScaling), customScaling);
                     final Vec3d rotation = block.getAsVec3d(block::getRotation);
 
                     try (OpenGL.With ignored1 = OpenGL.matrix(); OpenGL.With ignored2 = renderer.bindTexture(signalModel.getTextures())) {