+++ src/main/java/net/landofrails/landofsignals/render/block/TileDecoRender.java
@@ -29,7 +29,7 @@ public class TileDecoRender {
     private static final Map<String, OBJRender> cache = new HashMap<>();
     private static final Map<String, List<String>> groupCache = new HashMap<>();
 
-    private static void checkCache(String blockId, Map<String, ContentPackModel[]> models) {
+    public static void checkCache(String blockId, Map<String, ContentPackModel[]> models) {
         Optional<String> firstPath = models.keySet().stream().findFirst();
         if (!firstPath.isPresent())
             return;
@@ -82,6 +82,8 @@ public class TileDecoRender {
     }
 
     private static void renderBase(String blockId, TileDeco tile) {
+        Vec3d offset = tile.getOffset();
+        Vec3d customScaling = tile.getScaling();
 
         ContentPackDeco contentPackDeco = LOSBlocks.BLOCK_DECO.getContentpackDeco().get(blockId);
 
@@ -98,8 +100,8 @@ public class TileDecoRender {
 
             for (ContentPackModel baseModel : baseModels.getValue()) {
                 ContentPackBlock block = baseModel.getBlock();
-                Vec3d translate = block.getAsVec3d(block::getTranslation);
-                Vec3d scale = block.getAsVec3d(block::getScaling);
+                Vec3d translate = block.getAsVec3d(block::getTranslation).add(offset);
+                final Vec3d scale = Static.multiply(block.getAsVec3d(block::getScaling), customScaling);
                 Vec3d rotation = block.getAsVec3d(block::getRotation);
 
                 try (OpenGL.With ignored1 = OpenGL.matrix(); OpenGL.With ignored2 = renderer.bindTexture(baseModel.getTextures())) {