+++ src/main/java/net/landofrails/landofsignals/render/block/TileSignPartRender.java
@@ -29,7 +29,7 @@ public class TileSignPartRender {
     private static final Map<String, OBJRender> cache = new HashMap<>();
     private static final Map<String, List<String>> groupCache = new HashMap<>();
 
-    private static void checkCache(String blockId, Map<String, ContentPackModel[]> models) {
+    public static void checkCache(String blockId, Map<String, ContentPackModel[]> models) {
         Optional<String> firstPath = models.keySet().stream().findFirst();
         if (!firstPath.isPresent())
             return;
@@ -83,6 +83,8 @@ public class TileSignPartRender {
 
     private static void renderBase(String blockId, TileSignPart tile) {
 
+        Vec3d offset = tile.getOffset();
+        Vec3d customScaling = tile.getScaling();
         ContentPackSign contentPackSign = LOSBlocks.BLOCK_SIGN_PART.getContentpackSigns().get(blockId);
 
         if(contentPackSign == null) contentPackSign = LOSBlocks.BLOCK_SIGN_PART.getContentpackSigns().get(Static.MISSING);
@@ -97,8 +99,8 @@ public class TileSignPartRender {
 
             for (ContentPackModel baseModel : baseModels.getValue()) {
                 ContentPackBlock block = baseModel.getBlock();
-                Vec3d translate = block.getAsVec3d(block::getTranslation);
-                Vec3d scale = block.getAsVec3d(block::getScaling);
+                Vec3d translate = block.getAsVec3d(block::getTranslation).add(offset);
+                Vec3d scale = Static.multiply(block.getAsVec3d(block::getScaling), customScaling);
                 Vec3d rotation = block.getAsVec3d(block::getRotation);
 
                 try (OpenGL.With ignored1 = OpenGL.matrix(); OpenGL.With ignored2 = renderer.bindTexture(baseModel.getTextures())) {