+++ src/main/java/net/landofrails/landofsignals/render/block/TileSignalPartRender.java
@@ -12,6 +12,7 @@ import net.landofrails.landofsignals.LOSBlocks;
 import net.landofrails.landofsignals.LandOfSignals;
 import net.landofrails.landofsignals.render.item.ItemRenderException;
 import net.landofrails.landofsignals.tile.TileSignalPart;
+import net.landofrails.landofsignals.utils.HighlightingUtil;
 import net.landofrails.landofsignals.utils.Static;
 import org.lwjgl.opengl.GL11;
 
@@ -26,7 +27,6 @@ public class TileSignalPartRender {
     }
 
     private static final Map<String, OBJRender> cache = new HashMap<>();
-    protected static final Map<String, Boolean> cacheInfoOldContentPack = new HashMap<>();
 
     public static StandardModel render(final TileSignalPart tsp) {
         return new StandardModel().addCustom(() -> renderStuff(tsp));
@@ -52,12 +52,17 @@ public class TileSignalPartRender {
         }
         renderSignals(id, signal, tsp);
 
+        if(tsp.isHighlighting()){
+            HighlightingUtil.renderHighlighting();
+        }
+
     }
 
     @SuppressWarnings("java:S1134")
     private static void renderBase(String blockId, ContentPackSignal signal, TileSignalPart tile) {
 
         final Vec3d offset = tile.getOffset();
+        final Vec3d customScaling = tile.getScaling();
         final String base = signal.getBase();
         final String objPath = signal.getModel();
 
@@ -65,7 +70,6 @@ public class TileSignalPartRender {
             try {
                 String[] states = LOSBlocks.BLOCK_SIGNAL_PART.getAllStates(blockId);
                 cache.put(objPath, new OBJRender(new OBJModel(new Identifier(LandOfSignals.MODID, objPath), 0, Arrays.asList(states))));
-                cacheInfoOldContentPack.putIfAbsent(blockId, LOSBlocks.BLOCK_SIGNAL_PART.isOldContentPack(blockId));
             } catch (Exception e) {
                 throw new ItemRenderException("Error loading item model/renderer...", e);
             }
@@ -74,7 +78,10 @@ public class TileSignalPartRender {
 
         final float[] originalTranslate = signal.getTranslation();
         final Vec3d translate = new Vec3d(originalTranslate[0], originalTranslate[1], originalTranslate[2]).add(offset);
-        final float[] scale = signal.getScaling();
+        final float[] scale = signal.getScaling().clone();
+        scale[0] *= customScaling.x;
+        scale[1] *= customScaling.y;
+        scale[2] *= customScaling.z;
 
         try (OpenGL.With ignored1 = OpenGL.matrix(); OpenGL.With ignored2 = renderer.bindTexture(base)) {
 
@@ -98,6 +105,7 @@ public class TileSignalPartRender {
     private static void renderSignals(String blockId, ContentPackSignal signal, TileSignalPart tile) {
 
         final Vec3d offset = tile.getOffset();
+        final Vec3d customScaling = tile.getScaling();
         final String signalState = tile.getState();
         final String objPath = signal.getModel();
 
@@ -105,7 +113,6 @@ public class TileSignalPartRender {
             try {
                 String[] states = LOSBlocks.BLOCK_SIGNAL_PART.getAllStates(blockId);
                 cache.put(objPath, new OBJRender(new OBJModel(new Identifier(LandOfSignals.MODID, objPath), 0, Arrays.asList(states))));
-                cacheInfoOldContentPack.putIfAbsent(blockId, LOSBlocks.BLOCK_SIGNAL_PART.isOldContentPack(blockId));
             } catch (Exception e) {
                 throw new ItemRenderException("Error loading item model/renderer...", e);
             }
@@ -114,7 +121,10 @@ public class TileSignalPartRender {
 
         final float[] originalTranslate = signal.getTranslation();
         final Vec3d translate = new Vec3d(originalTranslate[0], originalTranslate[1], originalTranslate[2]).add(offset);
-        final float[] scale = signal.getScaling();
+        final float[] scale = signal.getScaling().clone();
+        scale[0] *= customScaling.x;
+        scale[1] *= customScaling.y;
+        scale[2] *= customScaling.z;
 
         try (OpenGL.With ignored1 = OpenGL.matrix(); OpenGL.With ignored2 = renderer.bindTexture(signalState)) {
 
@@ -130,6 +140,10 @@ public class TileSignalPartRender {
             tile.getWorld().breakBlock(tile.getPos());
 
         }
+    }
+
+    public static Map<String, OBJRender> cache(){
+        return cache;
     }
 
 }