+++ src/main/java/net/landofrails/landofsignals/render/item/ItemComplexSignalRender.java
@@ -30,27 +30,29 @@ public class ItemComplexSignalRender implements ItemRender.IItemModel {
     protected static final Map<String, OBJRender> cache = new HashMap<>();
     protected static final Map<String, List<String>> groupCache = new HashMap<>();
 
-    private static void checkCache(String itemId, Collection<ContentPackSignalGroup> groups, String identifier) {
+    private static final String SIGNAL_IDENTIFIER = "/signals/";
+
+    public static void checkCache(String itemId, Collection<ContentPackSignalGroup> groups) {
 
         // Get first group, get first state, get first model
         Optional<String> firstPath = groups.iterator().next().getStates().values().iterator().next().getModels().keySet().stream().findFirst();
 
         if (!firstPath.isPresent())
             return;
-        final String firstObjId = itemId + identifier + firstPath.get();
+        final String firstObjId = itemId + SIGNAL_IDENTIFIER + firstPath.get();
         if (cache.containsKey(firstObjId)) {
             return;
         }
 
         for (ContentPackSignalGroup group : groups) {
             for (ContentPackSignalState state : group.getStates().values()) {
-                checkCache(itemId, state.getModels(), identifier, false);
+                checkCache(itemId, state.getModels(), SIGNAL_IDENTIFIER, false);
             }
         }
 
     }
 
-    private static void checkCache(String itemId, Map<String, ContentPackModel[]> models, String identifier, boolean checkIfAlreadyExisting) {
+    public static void checkCache(String itemId, Map<String, ContentPackModel[]> models, String identifier, boolean checkIfAlreadyExisting) {
         if (checkIfAlreadyExisting) {
             Optional<String> firstPath = models.keySet().stream().findFirst();
             if (!firstPath.isPresent())
@@ -74,7 +76,7 @@ public class ItemComplexSignalRender implements ItemRender.IItemModel {
                 for (ContentPackModel signalModel : modelEntry.getValue()) {
                     String[] groups = signalModel.getObj_groups();
                     if (groups.length > 0) {
-                        Predicate<String> targetGroup = renderOBJGroup -> Arrays.stream(groups).anyMatch(renderOBJGroup::startsWith);
+                        Predicate<String> targetGroup = renderOBJGroup -> Arrays.stream(groups).filter(Objects::nonNull).anyMatch(renderOBJGroup::startsWith);
                         List<String> modes = renderer.model.groups().stream().filter(targetGroup)
                                 .collect(Collectors.toCollection(ArrayList::new));
                         String groupCacheId = objId + "@" + String.join("+", groups);
@@ -130,13 +132,6 @@ public class ItemComplexSignalRender implements ItemRender.IItemModel {
             final String path = baseModels.getKey();
 
             final String objId = itemId + "/base/" + path;
-            if (!cache.containsKey(objId)) {
-                try {
-                    cache.put(objId, new OBJRender(new OBJModel(new Identifier(LandOfSignals.MODID, path), 0)));
-                } catch (Exception e) {
-                    throw new ItemRenderException("Error loading item model/renderer...", e);
-                }
-            }
             final OBJRender renderer = cache.get(objId);
 
             for (ContentPackModel baseModel : baseModels.getValue()) {
@@ -173,7 +168,7 @@ public class ItemComplexSignalRender implements ItemRender.IItemModel {
     private static void renderSignals(String itemId, Map<String, String> itemGroupStates) {
         final Map<String, ContentPackSignalGroup> signalGroups = LOSBlocks.BLOCK_COMPLEX_SIGNAL.getContentpackComplexSignals().get(itemId).getSignals();
 
-        checkCache(itemId, signalGroups.values(), "/signals/");
+        checkCache(itemId, signalGroups.values());
 
         for (Map.Entry<String, ContentPackSignalGroup> signalGroup : signalGroups.entrySet()) {
 
@@ -183,15 +178,7 @@ public class ItemComplexSignalRender implements ItemRender.IItemModel {
 
                 final String path = signalModels.getKey();
 
-                final String objId = itemId + "/signals/" + path;
-                if (!cache.containsKey(objId)) {
-                    try {
-                        final Set<String> objTextures = LOSBlocks.BLOCK_COMPLEX_SIGNAL.getContentpackComplexSignals().get(itemId).getObjTextures().get(path);
-                        cache.put(objId, new OBJRender(new OBJModel(new Identifier(LandOfSignals.MODID, path), 0, objTextures)));
-                    } catch (final Exception e) {
-                        throw new ItemRenderException("Error loading item model/renderer...", e);
-                    }
-                }
+                final String objId = itemId + SIGNAL_IDENTIFIER + path;
                 final OBJRender renderer = cache.get(objId);
 
                 for (ContentPackModel signalModel : signalModels.getValue()) {