+++ src/main/java/net/landofrails/landofsignals/tile/TileSignalBox.java
@@ -13,6 +13,8 @@ import cam72cam.mod.util.Facing;
 import net.landofrails.landofsignals.LOSBlocks;
 import net.landofrails.landofsignals.LOSItems;
 import net.landofrails.landofsignals.packet.SignalBoxTileSignalPartPacket;
+import net.landofrails.landofsignals.utils.IManipulate;
+import net.landofrails.landofsignals.utils.LandOfSignalsUtils;
 import net.landofrails.landofsignals.utils.Static;
 
 import javax.annotation.Nullable;
@@ -20,7 +22,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @SuppressWarnings({"java:S116", "java:S1134", "java:S1133"})
-public class TileSignalBox extends BlockEntity {
+public class TileSignalBox extends BlockEntity implements IManipulate {
 
     @TagField("id")
     private String id;
@@ -31,6 +33,12 @@ public class TileSignalBox extends BlockEntity {
     @TagField("UuidTileTop")
     @Nullable
     private Vec3i tileSignalPartPos = Vec3i.ZERO;
+
+    @TagField("offset")
+    private Vec3d offset = Vec3d.ZERO;
+
+    @TagField("scaling")
+    private Vec3d scaling = new Vec3d(1,1,1);
 
     @TagField("signalType")
     @Nullable
@@ -58,6 +66,9 @@ public class TileSignalBox extends BlockEntity {
     private TileSignalPart tileSignalPart;
     private TileComplexSignal tileComplexSignal;
 
+    // client-only
+    private boolean highlighting = false;
+
     public TileSignalBox(String id, int rot) {
         this.id = id;
         this.blockRotate = rot;
@@ -78,8 +89,13 @@ public class TileSignalBox extends BlockEntity {
     }
 
     @Override
+    public IBoundingBox getRenderBoundingBox() {
+        return IBoundingBox.BLOCK.offset(getOffset());
+    }
+
+    @Override
     public boolean onClick(Player player, Player.Hand hand, Facing facing, Vec3d hit) {
-        if (!player.getHeldItem(hand).is(LOSItems.ITEM_CONNECTOR) && !player.isCrouching() && player.getWorld().isServer) {
+        if (!LandOfSignalsUtils.isLandOfSignalsItem(player.getHeldItem(Player.Hand.PRIMARY)) && !player.isCrouching() && player.getWorld().isServer) {
 
             if(signalType == null) {
                 refreshOldRedstoneVariables();
@@ -219,6 +235,11 @@ public class TileSignalBox extends BlockEntity {
     @Nullable
     public Byte getSignalType() {
         return this.signalType;
+    }
+
+    @Nullable
+    public Vec3i getTileSignalPartPos() {
+        return this.tileSignalPartPos;
     }
 
     /**
@@ -393,4 +414,52 @@ public class TileSignalBox extends BlockEntity {
         noRedstone = null;
     }
 
+    @Override
+    public void setOffset(Vec3d offset) {
+        this.offset = offset;
+    }
+
+    @Override
+    public Vec3d getOffset() {
+        return offset;
+    }
+
+    @Override
+    public void setRotation(int rotation) {
+        this.blockRotate = rotation;
+    }
+
+    @Override
+    public int getRotation() {
+        return getBlockRotate();
+    }
+
+    @Override
+    public void setScaling(Vec3d scaling) {
+        this.scaling = scaling;
+    }
+
+    @Override
+    public Vec3d getScaling() {
+        return scaling;
+    }
+
+    public void toggleHighlighting() {
+        this.highlighting = !this.highlighting;
+        if(signalType != null && 0 == signalType){
+            TileSignalPart signal = getWorld().getBlockEntity(tileSignalPartPos, TileSignalPart.class);
+            if(signal != null){
+                signal.setHighlighting(this.highlighting);
+            }
+        }else if(signalType != null && 1 == signalType){
+            TileComplexSignal signal = getWorld().getBlockEntity(tileSignalPartPos, TileComplexSignal.class);
+            if(signal != null){
+                signal.setHighlighting(this.highlighting);
+            }
+        }
+    }
+
+    public boolean isHighlighting(){
+        return this.highlighting;
+    }
 }