+++ src/main/java/net/landofrails/landofsignals/tile/TileSignalPart.java
@@ -45,6 +45,11 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
 
     @TagField("offset")
     private Vec3d offset = Vec3d.ZERO;
+    @TagField("scaling")
+    private Vec3d scaling = new Vec3d(1, 1, 1);
+
+    // client-only
+    private boolean highlighting = false;
 
     public TileSignalPart(final String id, final int rot) {
         this.blockRotate = rot;
@@ -53,7 +58,7 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
 
     @Override
     public boolean onClick(Player player, Player.Hand hand, Facing facing, Vec3d hit) {
-        if (player.isCrouching() || LandOfSignalsUtils.isLandOfSignalsItem(player.getHeldItem(hand))) {
+        if (player.isCrouching() || LandOfSignalsUtils.isLandOfSignalsItem(player.getHeldItem(Player.Hand.PRIMARY))) {
             return false;
         }
         if (!getWorld().isServer) {
@@ -76,7 +81,7 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
 
     @Override
     public IBoundingBox getBoundingBox() {
-        return IBoundingBox.BLOCK.offset(offset);
+        return IBoundingBox.BLOCK;
     }
 
     @Override
@@ -133,6 +138,16 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
         return getBlockRotate();
     }
 
+    @Override
+    public void setScaling(Vec3d scaling) {
+        this.scaling = scaling;
+    }
+
+    @Override
+    public Vec3d getScaling() {
+        return scaling;
+    }
+
     /**
      * server-only
      *
@@ -164,14 +179,14 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
 
         senderSignalStates.put(pos, state);
 
-        refreshSignals(true);
+        refreshSignals();
     }
 
     /**
      * server-only
      */
     public void updateSignals() {
-        refreshSignals(true);
+        refreshSignals();
     }
 
     /**
@@ -181,10 +196,10 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
      */
     public void removeSignal(Vec3i pos) {
         senderSignalStates.remove(pos);
-        refreshSignals(true);
+        refreshSignals();
     }
 
-    private void refreshSignals(boolean updateClients) {
+    private void refreshSignals() {
         boolean first = true;
         String lastState = null;
         for (String state : getOrderedStates()) {
@@ -195,9 +210,7 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
         }
         this.texturePath = lastState;
 
-        if (updateClients) {
-            new SignalUpdatePacket(getPos(), texturePath).sendToAll();
-        }
+        new SignalUpdatePacket(getPos(), texturePath).sendToAll();
     }
 
     @Override
@@ -255,4 +268,13 @@ public class TileSignalPart extends BlockEntity implements IManipulate {
 
         return foundActiveState && foundInactiveState;
     }
+
+    public void setHighlighting(boolean highlighting) {
+        this.highlighting = highlighting;
+    }
+
+    public boolean isHighlighting(){
+        return this.highlighting;
+    }
+
 }